factor out non-core utils
